+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Upgrade to CodeMirror 6 with autocomplete support in the CSS editor
  * Admin Toolbar: Cleanup the DOM properly after a drag ends
 
 ## 2.1.3 (July 2, 2022)